Add files via upload
A login form utilizes the credentials of a user, in order to authenticate their access. It generally consists of the typical username or email and password. But more fields can be added to improve the site's security. These can be in the form of a passcode, PIN number, or a secret phrase.